requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pymssql==2.3.2
psycopg2-binary==2.9.10
